+++ AdlsDotNetSDK/ADLSClient.cs
@@ -123,7 +123,11 @@ namespace Microsoft.Azure.DataLake.Store
         /// ContentEncoding for the compression, if not set then we use default contentencoding
         /// </summary>
         internal string ContentEncoding { get; set; } = null;
-        
+
+        private TimeSpan _perRequestTimeout = new TimeSpan(0, 0, 60);
+
+        private TimeSpan _perRequestTimeoutForEnumerateTrash = new TimeSpan(0, 0, 135);
+
         #endregion
 
         #region Constructors
@@ -354,6 +358,21 @@ namespace Microsoft.Azure.DataLake.Store
                 AccessToken = "Bearer " + accessToken;
             }
         }
+
+        /// <summary>
+        /// Sets the per request timeout. Highly recommended to set it after creating client. 
+        /// Not recommended to set it while requests are underway from a different thread.
+        /// </summary>
+        /// <param name="timeout"></param>
+        public void SetPerRequestTimeout(TimeSpan timeout)
+        {
+            _perRequestTimeout = timeout;
+        }
+
+        internal TimeSpan GetPerRequestTimeout()
+        {
+            return _perRequestTimeout;
+        }
         #endregion
 
         #region REST API 
@@ -376,7 +395,7 @@ namespace Microsoft.Azure.DataLake.Store
             }
             OperationResponse resp = new OperationResponse();
             bool result = await Core.MkdirsAsync(dirName, octalPermission, this,
-                new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+                new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in creating Directory {dirName}.");
@@ -437,7 +456,7 @@ namespace Microsoft.Azure.DataLake.Store
             // Pass getConsistentlength so that we get the updated length of stream
             DirectoryEntry diren = await Core
                 .GetFileStatusAsync(filename, UserGroupRepresentation.ObjectID, this,
-                    new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken, true).ConfigureAwait(false);
+                    new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken, true).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error opening a Read Stream for file {filename}.");
@@ -482,7 +501,7 @@ namespace Microsoft.Azure.DataLake.Store
             OperationResponse resp = new OperationResponse();
             //This is necessary to do to obtain the lease on the file
             await Core.AppendAsync(filename, leaseId, leaseId, SyncFlag.DATA, 0, null, -1, 0, this,
-                new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+                new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error trying to append to file {filename}.");
@@ -568,7 +587,7 @@ namespace Microsoft.Azure.DataLake.Store
             RetryPolicy policy = overwrite ? new ExponentialRetryPolicy() : (RetryPolicy)new NonIdempotentRetryPolicy();
 
             OperationResponse resp = new OperationResponse();
-            await Core.CreateAsync(filename, overwrite, octalPermission, leaseId, leaseId, createParent, SyncFlag.DATA, null, -1, 0, this, new RequestOptions(policy), resp, cancelToken).ConfigureAwait(false);
+            await Core.CreateAsync(filename, overwrite, octalPermission, leaseId, leaseId, createParent, SyncFlag.DATA, null, -1, 0, this, new RequestOptions(GetPerRequestTimeout(), policy), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in creating file {filename}.");
@@ -644,7 +663,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Cant delete the root");
             }
             OperationResponse resp = new OperationResponse();
-            bool isSucceded = await Core.DeleteAsync(path, false, this, new RequestOptions(new ExponentialRetryPolicy()),
+            bool isSucceded = await Core.DeleteAsync(path, false, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()),
                 resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
@@ -681,7 +700,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Cant delete the root");
             }
             OperationResponse resp = new OperationResponse();
-            bool isSucceded = await Core.DeleteAsync(path, true, this, new RequestOptions(new ExponentialRetryPolicy()),
+            bool isSucceded = await Core.DeleteAsync(path, true, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()),
                 resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
@@ -733,7 +752,7 @@ namespace Microsoft.Azure.DataLake.Store
                 }
             }
             OperationResponse resp = new OperationResponse();
-            bool isSucceeded = await Core.RenameAsync(path, destination, overwrite, this, new RequestOptions(new ExponentialRetryPolicy()),
+            bool isSucceeded = await Core.RenameAsync(path, destination, overwrite, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()),
                 resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
@@ -774,7 +793,7 @@ namespace Microsoft.Azure.DataLake.Store
             }
             OperationResponse resp = new OperationResponse();
             DirectoryEntry diren = await Core.GetFileStatusAsync(path, uid, this,
-                new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+                new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (diren == null)
             {
                 throw GetExceptionFromResponse(resp, $"Error in getting metadata for path {path}.");
@@ -808,7 +827,7 @@ namespace Microsoft.Azure.DataLake.Store
             }
             OperationResponse resp = new OperationResponse();
             await Core.ConcatAsync(destination, concatFiles, deleteSource, this,
-                new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+                new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in concating files {String.Join(",", concatFiles)} to destination {destination}");
@@ -962,7 +981,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.SetExpiryTimeAsync(path, eopt, expiryTime, this, new RequestOptions(new ExponentialRetryPolicy()),
+            await Core.SetExpiryTimeAsync(path, eopt, expiryTime, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()),
                 resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
@@ -995,7 +1014,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.CheckAccessSync(path, rwx, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.CheckAccessSync(path, rwx, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 if (resp.HttpStatus == HttpStatusCode.Forbidden || resp.HttpStatus == HttpStatusCode.Unauthorized)
@@ -1033,7 +1052,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.SetPermissionAsync(path, permission, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.SetPermissionAsync(path, permission, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in setting permission for path {path}.");
@@ -1065,7 +1084,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.ModifyAclEntriesAsync(path, aclSpec, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.ModifyAclEntriesAsync(path, aclSpec, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in modifying ACL entries {AclEntry.SerializeAcl(aclSpec, false)} for path {path}.");
@@ -1097,7 +1116,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.SetAclAsync(path, aclSpec, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.SetAclAsync(path, aclSpec, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in modifying ACL entries {AclEntry.SerializeAcl(aclSpec, false)} for path {path}.");
@@ -1142,7 +1161,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.SetOwnerAsync(path, owner, group, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.SetOwnerAsync(path, owner, group, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in setting owner or group for path {path}.");
@@ -1162,7 +1181,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.RemoveAclEntriesAsync(path, aclSpec, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.RemoveAclEntriesAsync(path, aclSpec, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in modifying ACL entries {AclEntry.SerializeAcl(aclSpec, true)} for path {path}.");
@@ -1193,7 +1212,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.RemoveAclAsync(path, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.RemoveAclAsync(path, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in removing all ACL entries for path {path}.");
@@ -1222,7 +1241,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.RemoveDefaultAclAsync(path, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            await Core.RemoveDefaultAclAsync(path, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in removing all default ACL entries for path {path}.");
@@ -1252,7 +1271,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null");
             }
             OperationResponse resp = new OperationResponse();
-            AclStatus status = await Core.GetAclStatusAsync(path, userIdFormat, this, new RequestOptions(new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+            AclStatus status = await Core.GetAclStatusAsync(path, userIdFormat, this, new RequestOptions(GetPerRequestTimeout(), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in getting ACL entries for path {path}.");
@@ -1314,8 +1333,8 @@ namespace Microsoft.Azure.DataLake.Store
             {
                 OperationResponse resp = new OperationResponse();
 
-                // TODO : pass the timeout programmatically instead of hard-coding here
-                TrashStatus trashstatus = await Core.EnumerateDeletedItemsAsync(hint, nextListAfter, numResults - numFound, this, new RequestOptions(null, new TimeSpan(0, 0, 135), new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
+                var requestTimeout = _perRequestTimeoutForEnumerateTrash < GetPerRequestTimeout() ? GetPerRequestTimeout() : _perRequestTimeoutForEnumerateTrash;
+                TrashStatus trashstatus = await Core.EnumerateDeletedItemsAsync(hint, nextListAfter, numResults - numFound, this, new RequestOptions(null, requestTimeout, new ExponentialRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
                 if (!resp.IsSuccessful)
                 {
                     if (!(resp.Ex is OperationCanceledException))
@@ -1384,7 +1403,7 @@ namespace Microsoft.Azure.DataLake.Store
         public virtual async Task RestoreDeletedItemsAsync(string pathOfFileToRestoreInTrash, string restoreDestination, string type, string restoreAction = "", CancellationToken cancelToken = default(CancellationToken))
         {
             OperationResponse resp = new OperationResponse();
-            await Core.RestoreDeletedItemsAsync(pathOfFileToRestoreInTrash, restoreDestination, type, restoreAction, this, new RequestOptions(), resp, cancelToken).ConfigureAwait(false);
+            await Core.RestoreDeletedItemsAsync(pathOfFileToRestoreInTrash, restoreDestination, type, restoreAction, this, new RequestOptions(GetPerRequestTimeout(), new NoRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in RestoreDeletedItemsAsync pathOfFileToRestoreInTrash:{pathOfFileToRestoreInTrash}, restoreDestination:{restoreDestination}, type:{type}, restoreAction:{restoreAction}");
@@ -1424,7 +1443,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null or empty");
             }
             OperationResponse resp = new OperationResponse();
-            await Core.ConcurrentAppendAsync(path, autoCreate, dataBytes, offset, length, this, new RequestOptions(), resp, cancelToken).ConfigureAwait(false);
+            await Core.ConcurrentAppendAsync(path, autoCreate, dataBytes, offset, length, this, new RequestOptions(GetPerRequestTimeout(), new NoRetryPolicy()), resp, cancelToken).ConfigureAwait(false);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in concurrent append for file {path}.");
@@ -1447,7 +1466,7 @@ namespace Microsoft.Azure.DataLake.Store
                 throw new ArgumentException("Path is null or empty");
             }
             OperationResponse resp = new OperationResponse();
-            Core.ConcurrentAppend(path, autoCreate, dataBytes, offset, length, this, new RequestOptions(), resp);
+            Core.ConcurrentAppend(path, autoCreate, dataBytes, offset, length, this, new RequestOptions(GetPerRequestTimeout(), new NoRetryPolicy()), resp);
             if (!resp.IsSuccessful)
             {
                 throw GetExceptionFromResponse(resp, $"Error in concurrent append for file {path}.");